+++ rust/sealevel/programs/ism/test-ism/Cargo.toml
@@ -5,6 +5,7 @@ edition = "2021"
 
 [features]
 no-entrypoint = []
+test-client = ["dep:solana-program-test", "dep:solana-sdk", "dep:hyperlane-test-utils"]
 
 [dependencies]
 account-utils = { path = "../../../libraries/account-utils" }
@@ -14,6 +15,10 @@ serializable-account-meta = { path = "../../../libraries/serializable-account-me
 hyperlane-core = { path = "../../../hyperlane-core" }
 hyperlane-sealevel-mailbox = { path = "../../mailbox", features = ["no-entrypoint"] }
 hyperlane-sealevel-interchain-security-module-interface = { path = "../../../libraries/interchain-security-module-interface" }
+
+solana-program-test = { version = "=1.14.13", optional = true }
+solana-sdk = { version = "=1.14.13", optional = true }
+hyperlane-test-utils = { path = "../../../libraries/test-utils", optional = true }
 
 [lib]
 crate-type = ["cdylib", "lib"]
