+++ rust/sealevel/programs/mailbox/Cargo.toml
@@ -36,7 +36,7 @@ solana-program-test = "=1.14.13"
 solana-sdk = "=1.14.13"
 hyperlane-test-utils = { path = "../../libraries/test-utils" }
 hyperlane-sealevel-ism-rubber-stamp = { path = "../ism/rubber-stamp" }
-hyperlane-sealevel-test-ism = { path = "../ism/test-ism" }
+hyperlane-sealevel-test-ism = { path = "../ism/test-ism", features = ["test-client"] }
 hyperlane-sealevel-test-send-receiver = { path = "../test-send-receiver", features = ["test-client"] }
 
 [lib]
