+++ rust/sealevel/programs/test-send-receiver/Cargo.toml
@@ -5,7 +5,7 @@ edition = "2021"
 
 [features]
 no-entrypoint = []
-test-client = ["dep:solana-program-test", "dep:solana-sdk", "dep:hyperlane-test-utils"]
+test-client = ["dep:solana-program-test", "dep:solana-sdk", "dep:hyperlane-test-utils", "dep:spl-noop"]
 
 [dependencies]
 account-utils = { path = "../../libraries/account-utils" }
@@ -18,6 +18,14 @@ hyperlane-sealevel-message-recipient-interface = { path = "../../libraries/messa
 solana-program-test = { version = "=1.14.13", optional = true }
 solana-sdk = { version = "=1.14.13", optional = true }
 hyperlane-test-utils = { path = "../../libraries/test-utils", optional = true }
+
+[dependencies.spl-noop]
+version = "=0.1.3"
+features = ["no-entrypoint"]
+git = "https://github.com/tkporter/eclipse-program-library"
+branch = "trevor/steven/eclipse-1.14.13/with-tlv-lib"
+optional = true
+
 
 [lib]
 crate-type = ["cdylib", "lib"]
