+++ rust/sealevel/programs/test-send-receiver/src/test_client.rs
@@ -5,9 +5,13 @@ use solana_program::{
     system_program,
 };
 use solana_program_test::*;
-use solana_sdk::{signature::Signer, signer::keypair::Keypair};
+use solana_sdk::{signature::Signature, signature::Signer, signer::keypair::Keypair};
 
-use hyperlane_test_utils::process_instruction;
+use hyperlane_sealevel_mailbox::{
+    instruction::OutboxDispatch, mailbox_dispatched_message_pda_seeds,
+    mailbox_message_dispatch_authority_pda_seeds,
+};
+use hyperlane_test_utils::{process_instruction, MailboxAccounts};
 
 use crate::{
     id,
@@ -52,7 +56,9 @@ impl TestSendReceiverTestClient {
             &self.payer,
             &[&self.payer],
         )
-        .await
+        .await?;
+
+        Ok(())
     }
 
     pub async fn set_ism(
@@ -82,7 +88,9 @@ impl TestSendReceiverTestClient {
             &self.payer,
             &[&self.payer],
         )
-        .await
+        .await?;
+
+        Ok(())
     }
 
     pub async fn set_fail_handle(&mut self, fail_handle: bool) -> Result<(), BanksClientError> {
@@ -108,7 +116,69 @@ impl TestSendReceiverTestClient {
             &self.payer,
             &[&self.payer],
         )
-        .await
+        .await?;
+
+        Ok(())
+    }
+
+    pub async fn dispatch(
+        &mut self,
+        mailbox_accounts: &MailboxAccounts,
+        outbox_dispatch: OutboxDispatch,
+    ) -> Result<(Signature, Keypair, Pubkey), BanksClientError> {
+        let program_id = id();
+
+        let unique_message_account_keypair = Keypair::new();
+
+        let (dispatch_authority_key, _expected_dispatch_authority_bump) =
+            Self::get_dispatch_authority();
+
+        let (dispatched_message_account_key, _dispatched_message_bump) =
+            Pubkey::find_program_address(
+                mailbox_dispatched_message_pda_seeds!(&unique_message_account_keypair.pubkey()),
+                &mailbox_accounts.program,
+            );
+
+        let instruction = Instruction {
+            program_id,
+            data: TestSendReceiverInstruction::Dispatch(outbox_dispatch)
+                .try_to_vec()
+                .unwrap(),
+            accounts: vec![
+                // 0. [executable] The Mailbox program.
+                // And now the accounts expected by the Mailbox's OutboxDispatch instruction:
+                // 1. [writeable] Outbox PDA.
+                // 2. [] This program's dispatch authority.
+                // 3. [executable] System program.
+                // 4. [executable] SPL Noop program.
+                // 5. [signer] Payer.
+                // 6. [signer] Unique message account.
+                // 7. [writeable] Dispatched message PDA. An empty message PDA relating to the seeds
+                //    `mailbox_dispatched_message_pda_seeds` where the message contents will be stored.
+                AccountMeta::new_readonly(mailbox_accounts.program, false),
+                AccountMeta::new(mailbox_accounts.outbox, false),
+                AccountMeta::new_readonly(dispatch_authority_key, false),
+                AccountMeta::new_readonly(system_program::id(), false),
+                AccountMeta::new_readonly(spl_noop::id(), false),
+                AccountMeta::new(self.payer.pubkey(), true),
+                AccountMeta::new(unique_message_account_keypair.pubkey(), true),
+                AccountMeta::new(dispatched_message_account_key, false),
+            ],
+        };
+
+        let tx_signature = process_instruction(
+            &mut self.banks_client,
+            instruction,
+            &self.payer,
+            &[&self.payer, &unique_message_account_keypair],
+        )
+        .await?;
+
+        Ok((
+            tx_signature,
+            unique_message_account_keypair,
+            dispatched_message_account_key,
+        ))
     }
 
     fn get_storage_pda_key() -> Pubkey {
@@ -116,6 +186,11 @@ impl TestSendReceiverTestClient {
         let (storage_pda_key, _storage_pda_bump) =
             Pubkey::find_program_address(test_send_receiver_storage_pda_seeds!(), &program_id);
         storage_pda_key
+    }
+
+    fn get_dispatch_authority() -> (Pubkey, u8) {
+        let program_id = id();
+        Pubkey::find_program_address(mailbox_message_dispatch_authority_pda_seeds!(), &program_id)
     }
 
     pub fn id(&self) -> Pubkey {
